Move all db code into one subdirectory
To further decouple db specific from mediator code.

OHIE-376
